♻️ refactor (footer): Creer component pour Newsletter
