+++ src/components/NewsletterFooter.jsx
